+++ priv/repo/migrations/20231004113255_create_tickets.exs
@@ -5,7 +5,7 @@ defmodule ElixirConfAfrica.Repo.Migrations.CreateTickets do
     create table(:tickets) do
       add :type, :string, null: false
       add :description, :string, null: false
-      add :price, :integer, null: false
+      add :price, :decimal, null: false
       add :ticket_number, :integer, null: false
       add :event_id, references(:events, on_delete: :nothing)
